Move common and constants to ms2lda to prevent import errors
