+++ ms2lda/common.py
