+++ ms2lda/constants.py
